Separated helpers into their old file
